refactor(cypress): unify commands into cy.loginToMerchantCenter and omit cy.origin usage
