+++ cypress/e2e/playground/state-machines.cy.ts
@@ -6,8 +6,10 @@ import {
 
 describe('State machines', () => {
   beforeEach(() => {
-    cy.loginByForm({ entryPointUriPath: ENTRY_POINT_APP_KIT_PLAYGROUND });
-    cy.visit(URL_APP_KIT_PLAYGROUND);
+    cy.loginToMerchantCenter({
+      entryPointUriPath: ENTRY_POINT_APP_KIT_PLAYGROUND,
+      initialRoute: URL_APP_KIT_PLAYGROUND,
+    });
   });
   it('should render list view', () => {
     cy.get('main').within(() => {
